机器人 + attributeLevelUpReqPb.money:int -> long

+++ common-pb/src/main/java/common/pb/pb/AttributeLevelUpReqPb.java
@@ -29,5 +29,5 @@ public class AttributeLevelUpReqPb {
     int targetLevel;
     /** 升级费用 */
     @NotNull
-    int moneyCost;
+    long moneyCost;
 }
